+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGridInternal/MergePropertyDescriptor.cs
@@ -14,7 +14,7 @@ using System.Runtime.Serialization.Formatters.Binary;
 
 namespace System.Windows.Forms.PropertyGridInternal
 {
-    internal class MergePropertyDescriptor : PropertyDescriptor
+    internal partial class MergePropertyDescriptor : PropertyDescriptor
     {
         private readonly PropertyDescriptor[] descriptors;
 
@@ -436,204 +436,6 @@ namespace System.Windows.Forms.PropertyGridInternal
             }
 
             return true;
-        }
-
-        private class MultiMergeCollection : ICollection
-        {
-            private object[] items;
-            private bool locked;
-
-            public MultiMergeCollection(ICollection original)
-            {
-                SetItems(original);
-            }
-
-            /// <summary>
-            ///  Retrieves the number of items.
-            /// </summary>
-            public int Count
-            {
-                get
-                {
-                    if (items is not null)
-                    {
-                        return items.Length;
-                    }
-                    else
-                    {
-                        return 0;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Prevents the contents of the collection from being re-initialized;
-            /// </summary>
-            public bool Locked
-            {
-                get
-                {
-                    return locked;
-                }
-                set
-                {
-                    locked = value;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public void CopyTo(Array array, int index)
-            {
-                if (items is null)
-                {
-                    return;
-                }
-
-                Array.Copy(items, 0, array, index, items.Length);
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (items is not null)
-                {
-                    return items.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<object>().GetEnumerator();
-                }
-            }
-
-            /// <summary>
-            ///  Ensures that the new collection equals the exisitng one.
-            ///  Otherwise, it wipes out the contents of the new collection.
-            /// </summary>
-            public bool MergeCollection(ICollection newCollection)
-            {
-                if (locked)
-                {
-                    return true;
-                }
-
-                if (items.Length != newCollection.Count)
-                {
-                    items = Array.Empty<object>();
-                    return false;
-                }
-
-                object[] newItems = new object[newCollection.Count];
-                newCollection.CopyTo(newItems, 0);
-                for (int i = 0; i < newItems.Length; i++)
-                {
-                    if (((newItems[i] is null) != (items[i] is null)) ||
-                        (items[i] is not null && !items[i].Equals(newItems[i])))
-                    {
-                        items = Array.Empty<object>();
-                        return false;
-                    }
-                }
-
-                return true;
-            }
-
-            public void SetItems(ICollection collection)
-            {
-                if (locked)
-                {
-                    return;
-                }
-
-                items = new object[collection.Count];
-                collection.CopyTo(items, 0);
-            }
-        }
-
-        private class MergedAttributeCollection : AttributeCollection
-        {
-            private readonly MergePropertyDescriptor owner;
-
-            private AttributeCollection[] attributeCollections;
-            private IDictionary foundAttributes;
-
-            public MergedAttributeCollection(MergePropertyDescriptor owner) : base((Attribute[])null)
-            {
-                this.owner = owner;
-            }
-
-            public override Attribute this[Type attributeType]
-            {
-                get
-                {
-                    return GetCommonAttribute(attributeType);
-                }
-            }
-
-            private Attribute GetCommonAttribute(Type attributeType)
-            {
-                if (attributeCollections is null)
-                {
-                    attributeCollections = new AttributeCollection[owner.descriptors.Length];
-                    for (int i = 0; i < owner.descriptors.Length; i++)
-                    {
-                        attributeCollections[i] = owner.descriptors[i].Attributes;
-                    }
-                }
-
-                if (attributeCollections.Length == 0)
-                {
-                    return GetDefaultAttribute(attributeType);
-                }
-
-                Attribute value;
-                if (foundAttributes is not null)
-                {
-                    value = foundAttributes[attributeType] as Attribute;
-                    if (value is not null)
-                    {
-                        return value;
-                    }
-                }
-
-                value = attributeCollections[0][attributeType];
-
-                if (value is null)
-                {
-                    return null;
-                }
-
-                for (int i = 1; i < attributeCollections.Length; i++)
-                {
-                    Attribute newValue = attributeCollections[i][attributeType];
-                    if (!value.Equals(newValue))
-                    {
-                        value = GetDefaultAttribute(attributeType);
-                        break;
-                    }
-                }
-
-                if (foundAttributes is null)
-                {
-                    foundAttributes = new Hashtable();
-                }
-
-                foundAttributes[attributeType] = value;
-                return value;
-            }
         }
     }
 }
